Create IdenticalArrayMinSwap.java
ou are given two input arrays of same size n, which contains the same elements but placed in a different order. Elements present in the first or second array should be distinct. You have to make the second array same as the first array using a minimum number of swaps.
Note: We can only make changes in the second array.