Create FaultyKeyword.java
Some of the keys on your keyboard are faulty. When you press a key, it may register more than once. That means if you want to type "code", the keyboard may interpret it as "code", "cccoddee" or "coode" or "codeeeee", etc. So, by pressing the keys C, O, D, and E in that order, you may get any of the above-mentioned words.
Sample Input 1
2
code
cooodeee
hello
hheeeloo
Sample Output 1
true
false