Time Complexity: O(4 ^n )
Given an integer n, using phone keypad find out and print all the possible strings that can be made using digits of input n.
Note : The order of strings are not important. Just print different strings in new lines.
Input Format :
Time Complexity: O(4^n), where n is a number of digits in the input number.
Each digit of a number has 3 or 4 alphabets, thus it can be said that each digit has 4 alphabets as
options. If there are n digits then there are 4 options for the first digit and for each alphabet of
the first digit there are 4 options in the second digit, i.e for every recursion 4 more recursions are
called (if it does not match the base case). Therefore the time complexity is O(4n).