Time Complexity: O(2 ^N )
Given an array A of size n and an integer K, return all subsets of A which sum to K.
Sample Input :
9 
5 12 3 17 1 18 15 3 17 
6
Sample Output :
3 3
5 1

Time Complexity: O(2^N), where N is the size of the input array. Since, for every index i, two
recursive cases originate, hence the time complexity is O(2N).