Create PalindromePartitioning.java
Given a string s, partition s such that every substring of the partition is a palindrome. Return all possible palindrome partitioning of s.

A palindrome string is a string that reads the same backward as forward.

 

Example 1:

Input: s = "aab"
Output: [["a","a","b"],["aa","b"]]
Example 2:

Input: s = "a"
Output: [["a"]]

Time Complexity : O(N⋅2^ N ), where N is the length of string s. This is the worst-case time complexity when all the possible substrings are palindrome.
Example, if = s = aaa, the recursive tree can be illustrated as follows:



Hence, there could be 2^{N}2 N
  possible substrings in the worst case. For each substring, it takes \mathcal{O}(N)O(N) time to generate substring and determine if it a palindrome or not. This gives us time complexity as \mathcal{O}(N \cdot 2^{N})O(N⋅2 N )

Space Complexity: \mathcal{O}(N)O(N), where NN is the length of the string ss. This space will be used to store the recursion stack. For s = aaa, the maximum depth of the recursive call stack is 3 which is equivalent to NN.