Create saveTheVillage.java
There is a small village present near the city, which is causing a lot of problems to the people present in the city. So, the committee head (Shyam) of the city, wants to destroy the village himself. Ram is the head of the village, so, he ordered his soldiers to fight for the village.
But Ram is confused whether they should confront Shyam or evacuate the villagers including the soldiers. So, you have to help him out in this decision.
Shyam has a strength of Z and is confident that he will succeed. Ram has N soldiers under his command numbered 1 through N. Power of i-th soldier is denoted by Ai. When a soldier attacks Shyam, his strength gets reduced by the corresponding power of the soldier. However, every action has a reaction so the power of the soldier also gets halved i.e. Ai changes to [Ai/2]. Each soldier may attack any number of times (including 0).Shyam is defeated if his strength is reduced to 0 or less.
Find the minimum number of times the soldiers need to attack so that the village is saved.

Sample Input:
6 30
1 20 3 4 2 10
Sample Output:
2