Create KLargestSums.java
You are given a N-sized array and two integers m and k. You have to check if the product of the k largest sums contiguous subarrays is greater than m.

Sample Input:
4
8 -4 -2 7
3 659
Sample Output:False

Explanation:
The 3 largest sums contiguous subarrays are: {8}, {7} and {8, -4, -2, 7}. Product of their sums is : 8*7*9=504 and it is less than 659.