Create flattenAMultilevelLL.java
You are given a linked list in which each node has two pointers: next and child. Each child pointer will have its own next and child pointer. This results in a multilevel data structure. For example, let us take following multi level linked list:

You have to flatten this multilevel linked list and form a singly linked list such that all the nodes in the first level should come first, then nodes of second level and so on.

Problem Description: You are given a linked list, where each node has a child pointer
associated with it, which points to a separate linked list. Each child node can further have its own
next and child pointer. This will result in a multi-level linked list. You have to flatten this linked
list into a single list such that all the nodes in the first level should come first, then the nodes of
the second level and so on. You have to return the head of the flatten linked list.
How to approach?
We have to flatten the given multi-level linked list such that first level comes first, then the
second level and so on. So, we will start processing our nodes from the first level one by one,
and if a node has a child then we will place that node at the end of the first level. After
processing all the nodes of the first level we will repeat the process again on the nodes which are
appended at the end of the first level and we will continue further.
We will take a pointer, let us say “hp” as the head of the first level of the linked list and a “tail”
pointer which will point to the end of the first level of the linked list.
If the current node has a child, then append this new child to the tail, find the last node of the
new child list appended and update the tail and move to the next node. Follow the same
procedure until we reach the last node.


Sample Input 1:
5
10 Y 4 N Y 20 Y 2 N n Y 13 Y 16 Y 3 N N N N 5 N 12 N 7 Y 17 y 9 y 19 n y 15 n n y 8 n n y 6 n n 11 N
Sample Output 1:
10 5 12 7 11 4 20 13 17 6 2 16 9 8 3 19 15