Create cloneALL.java
A singly linked list is given to you with N nodes where each node is pointing to its next node and some of its nodes are pointing to the arbitrary nodes. You are given with M arbitrary pointers, where you have M pairs denoting that two nodes a and b are connected i.e. a's arbitrary pointer is pointing towards b (a->arb=b). You have to write a function to clone the original linked list given to you and return its head. We will validate whether the linked list is a copy of the original linked list or not.
Note: If there is any node whose arbitrary pointer is not given then its by default null.


Approach 2:
To optimize the space complexity, we should think of another approach. To reduce the additional
space used we will proceed as follows:
1. We will create a copy of node1 and insert it between node1 and node2 in the original
linked list. Similarly, create a copy of node2 and insert it between node2 and node3 and
so on. Insert the copy of the Nth node after the Nth node.
2. Now, original->next node is the copy of the original node and original->arbitrary->next is
the copy of arbitrary.
Original->next->arbitrary=original->arbitrary->next
Traverse two nodes in the next iteration.
3. Now, restore the original and copy linked lists in a single traversal.
Original->next=original->next->next
copy->next=copy->next->next
4. Now, make original->next as NULL and return the cloned linked list.
This approach has a time complexity of O(n) and a constant space complexity.