Create flattenAMultilevelLL(Depthwise).java
Problem Description: You are given a linked list, where each node has a child pointer
associated with it, which points to a separate linked list. Each child node can further have its own
next and child pointer. This will result in a multi-level linked list. You have to flatten this linked
list into a single list such that the child node is printed before the next node for every node.
You have to return the head of the flatten linked list.
How to approach?
We will proceed with this problem using recursion, we will recursively flatten the linked list and
while flattening we will keep track of last visited node so that the next node can be linked after
that. We will store the next node of the current node. So, we will get the next list using the next
pointer and then we will recursively flatten the next list and attach itself after the last visited
node. If a down list exists, process it first, add down list as next of current node; else if next
exists, add it after the next of last added node.

Sample Input:
10
10 Y 4 N Y 20 Y 2 N n Y 13 Y 16 Y 3 N N N N 5 N 12 N 7 Y 17 y 9 y 19 n y 15 n n y 8 n n y 6 n n 11 N 12 N 13 N 14 N 15 N 16 N
Sample Output:
10 4 20 2 13 16 3 5 12 7 17 9 19 15 8 6 11 12 13 14 15 16