Create loopDetector.java
Detect a cycle in a linked list if there is any and return that node. If there is no cycle, return NULL. Note that the head pointer may be 'NULL' if the list is empty.
Input Format :
Line 1 : Linked List (Random)

Line 2 : Integer which may or may not be in linked list. If there be assume circular linked list on that node.

Output Format :
Integer if there is cycle in linked list otherwise 'null'.

Sample Input :
 1 2 3 4 5 6 7 8 -1
 5
Sample Output :
 5