time o(n) && space o(n)
You are given a histogram in the form of an array of size N. Each element in the array represents the height of the histogram's bar. You may assume that the width of each bar is 1 unit. Find out the area of the largest rectangle that can fit inside the histogram.

Example 1:
Input: heights = [2,1,5,6,2,3]
Output: 10
Explanation: The above is a histogram where width of each bar is 1.
The largest rectangle is shown in the red area, which has an area = 10 units.

Example 2:
Input: heights = [2,4]
Output: 4