TC:O(N), SC:ON)
You are given a histogram in the form of an array of size N. Each element in the array represents the height of the histogram's bar. You may assume that the width of each bar is 1 unit. Find out the area of the largest rectangle that can fit inside the histogram.

Sample Input 1
2
3 5
Sample Output 1
6
Sample Input 2
6
2 1 5 6 2 3
Sample Output 2
10