TC:O(N), SC:O(1)
There is a circle and there are N petrol pumps on that circle.
You have the following 2 sets of data:
1. Amount of petrol that each petrol pump has.
2. The distance of the next petrol pump from that petrol pump.
Write a function to find the first point from where the truck should start so that it could complete the whole circle. Print -1 if there is no starting point for the truck.
Note: There is no capacity limit of the truck and truck stops at each petrol pump and 1 liter of petrol can help the truck to move 1 unit.

Sample Input:
5
1 3 2 4 6 2 5 3 10 3
Sample output:
2
Explanation:
There are 4 petrol pumps with amount of petrol and distance to next petrol pump value pairs as {1, 3}, {2, 4}, {6, 2}, {5, 3} and {10, 3}. The first point from where truck can make a circular tour is 3rd petrol pump. Output in this case is 2 (index of 2nd petrol pump)