TC:O(nlogn+n+m), SC:O(min(n,m))
You have been given two integer arrays/list(ARR1 and ARR2) of size N and M, respectively. You need to print their intersection; An intersection for this problem can be defined when both the arrays/lists contain a particular value or to put it in other words, when there is a common value that exists in both the arrays/lists.
Note :
Input arrays/lists can contain duplicate elements.

The intersection elements printed would be in the order they appear in the first sorted array/list(ARR1).

Approach 1: The first approach we can think is that we can run 2 loops. In the outer loop pick
one element of array 1 and then with the help of an inner loop check whether the same element
exists in array 2 or not. If the same element is found then, we found an intersection so we print it,
and to handle the case of duplicates, we can replace this element with minus infinity.
Pseudo Code for this approach:
Function intersection:
 For i = 0 to i less than size1:
 For j = 0 to j less than size2:
 If arr1[i] equal to arr2[j]:
 Print(arr2[j])
 arr2[j]=minus infinity
 break


Approach 2: A better solution for this problem is to sort both the arrays. Now, we have to find
the intersection of 2 sorted arrays:
1) Use two index variables i and j, initialize them as i = 0, j = 0
2) If arr1[i] is smaller than arr2[j] then increment i.
3) If arr1[i] is greater than arr2[j] then increment j.
4) If both are same then print any of them and increment both i and j.
Time Complexity for this approach: Time complexity for sorting both arrays will be
O(size1*log(size1) + size2*log(size2)) and then finding the intersection will have a time
complexity of O(size1+size2).