TC:O(N), SC:O(N), where N is the size of i/p array
You are given a N-sized array of integers and a value k. In the given array you have to find whether this array can be divided into pairs such that the sum of every pair is divisible by k.
Sample Input:
8 
1 5 0 4 4 3 7 0
8
Sample Output:
true
Explanation:
We can divide array into (1, 7), (5, 3), (4, 4) and (0,0). Sum of all of these pairs is a multiple of 8.