Create Triangle-Array.java
Harshit gave an array to Aahad and asked to make a triangle and calculate the topmost element of this array. For each layer, the element is the sum of two number below them and last layer is the given array.
For Example, for the array {1, 2, 3, 4}, the triangle will be:
        20
      8   12
   3    5    7
1    2    3     4
The Last layer is an array and each element is the sum of two numbers below. Your task is to find the top element of the array. As the element is very large, print answer modulo 1000000007.
Sample Input1:
5
1 2 3 4 5
Sample Output1:
48
Explaination
In the Sample case, the triangle formed will be:
      [48]
    [20, 28]
   [8, 12, 16]
  [3, 5, 7, 9]
 [1, 2, 3, 4, 5]
and the top most element is 48.