Create Bank-Account.java
Using a class we need to handle the bank account of the customer. The customer can perform following task-
1. Deposit - to deposit amount into the account.
2. Withdraw - to withdraw an amount from the account.
3. Add Interest- to add interest balance to the account (integer part only).
(After adding interest balance is increased by interest times).
Complete the given functions to perform the desired task.
Input Format:
Line 1: Initial balance and interest. 

Line 2: Integer x. (1 for deposit, 2 for withdrawal, 3 to add interest, -1 to exit)

Line 3 : The amount to be deposited if x = 1. The amount to withdraw if x = 2. Empty if x = 3.

Line 4, 5 : Same as lines 2 and 3.
.
.
.
Perform the operations till -1 is encountered.
Output format:
Print the balance of account when -1 is encountered.
Sample Input 1 :
1000 0.10
1
400
2
250
3
-1
Sample Output 1 :
1265
Sample Input 2 :
1000 0.10
1
400
3
-1
Sample Output 2 :
1540