Create Stack-With-Middle-Operation.java
1. You have to design a Stack class in which the user can access and delete the middle element in O(1) running time complexity.
2. The data members of the Stack class must be public.
Implement the following public functions:
1. Constructor:
 This function will initialize the data members.
2. push:
This function should take one argument of type int and has return type void. This function should insert an element in the stack. Time complexity should be O(1).
3. pop:
This function takes no input arguments and has return type int. This should removes the last element which is entered and return that element as an answer. Return 0 if stack is empty.
Time complexity should be O(1).    
4. middle:
This function takes no input arguments and has return type int. This should return the middle element of stack. Return 0, if the stack is empty. Time Complexity: O(1).
5. removeMiddle:
This function takes no input arguments and has a return type int. This should remove and return the data of middle element of stack. Return 0, if the stack is empty.Time Complexity: O(1)
6. print:
This function takes no input arguments and has void return type. This should print all the elements of the stack. Time Complexity: O(n) 
Note: If there are even number of elements in the stack, we should consider the first middle element as the middle one. For example, in case of {1,2,3,4}; 2 is considered as the middle element.
Input Format:
The first and only line of input contains series of integers, terminated by -1. The integers represent choice. 
For choice equal to 1, push function is called. When choice is 1, it is followed by another integer that denotes element to be inserted in the stack.
For choice equal to 2, pop function is called.
For choice equal to 3, middle function is called.
For choice equal to 4, removeMiddle function is called.
For choice equal to 5, print function is called.
Constraints:
Time Limit: 1 second
Output format:
The output lines are formatted according to the problem description.
Sample Input 1:
1 1 1 2 1 3 1 4 1 5 3 -1
Sample Output 1:
3