TC:O(H),  SC:O(1)
You are given a Binary Search Tree containing positive integer values greater than 0. You have to check whether this tree contains a dead-end or not. The dead-end is a leaf node, after which we are not able to insert any node. That is, a dead end is said to exist in a BST if there is a leaf node to which a new child node cannot be added without violating the properties of a BST.
Input Format:
The first line of input contains the number of nodes(N) to be inserted in the BST. The following line of input contains N space-separated elements to be inserted in the BST.
Output Format:
 The only output line contains "yes" if the dead-end is present in the BST otherwise "no" is printed.
Constraints:
1<=Numberof nodes<=1000
Sample Input:
6
8 7 10 9 13 2
Sample Output:
 yes
Explanation:
The BST is:
                         8
                        / \
                       7   10
                      /    / \
                     2    9  13

We can't insert any element at node 9.  Hence, 9 is a dead-end.