Create Find-The-closest-Element.java
You are given a BST in the level order form and a target node K. You have to find a node having the minimum absolute difference with target value K and print the value of that minimum absolute difference.
Note: You have to return the value of the node having the minimum absolute difference with target value K, printing the minimum absolute difference is handled automatically.
Input Format
Line 1: Nodes in the level order form (separated by space). If any node does not have left or right child, take -1 in its place
Line 2: Value of the target node K
Output Format:
The minimum absolute difference of a node with given target value K.
Constraints:
1<=Number of nodes<=1000
Sample Input:
9 4 17 3 6 -1 22 -1 -1 5 7 20 -1 -1 -1 -1 -1 -1 -1
18
Sample Output:
1
Explanation:
Here K=18, The node that has value nearest to K is 17 so it's difference with k is 1 is the answer.