Create Sum-Of-elements-<or=to-Kth-Smallest-Element.java
You are given an integer K and elements of a binary search tree in level order form. You have to return the sum of elements which are smaller than or equal to Kth smallest element of the tree.
Input format :
The first line of input contains an integer that denotes the value of K. The following line contains an integer that represents the size of array. Let the size of array be denoted by the symbol n. n elements are required to represent binary search tree in level order form. In level order form, an element at an index i is parent to elements at index 2*i+1 and 2*i+2. The index of left child and right child are 2*i+1 and 2*i+2 respectively. So, the following line contains n space separated integers, that represents the elements in level order form (If any node does not have left or right child, take -1 in its place).
Output Format :
Elements/ nodes are printed using inorder traversal. Each node is printed in a new line. The nodes are printed in the following format:
Data of left child “ => ” Data of node “[” Data of parent “]” “=>” Data of right child
If left child or right child is null, then it is to be represented by “.”. If the parent is null, then it is to be represented by “null”.
Sample Input 1:
3
15
50 25 75 12 37 62 87 -1 -1 -1 -1 -1 -1 -1 -1
Sample Output 1:
74